+++ kernel/examples/CMakeLists.txt
@@ -6,7 +6,7 @@ ADD_SUBDIRECTORY(example_3)
 ADD_SUBDIRECTORY(example_4)
 ADD_SUBDIRECTORY(example_5)
 ADD_SUBDIRECTORY(example_6)
-ADD_SUBDIRECTORY(example_7)
+#ADD_SUBDIRECTORY(example_7)
