+++ kernel/examples/example_2/exe.cpp
@@ -1,5 +1,5 @@
-#include "kernel/executors/thread_pool.h"
-#include "kernel/executors/stoppable_task.h"
+#include "kernel/parallel/threading/thread_pool.h"
+#include "kernel/parallel/threading/stoppable_task.h"
 #include "kernel/base/types.h"
 
 #include <chrono>
