+++ kernel/examples/example_3/exe.cpp
@@ -3,10 +3,10 @@
   * TODO: Review/finalize implementation
   **/
 
-#include "kernel/executors/thread_pool.h"
-#include "kernel/executors/simple_task.h"
 #include "kernel/base/types.h"
-#include "kernel/utilities/read_write_lock.h"
+#include "kernel/parallel/threading/thread_pool.h"
+#include "kernel/parallel/threading/simple_task.h"
+#include "kernel/parallel/threading/read_write_lock.h"
 
 #include <thread>
 #include <vector>
