+++ kernel/examples/example_5/exe.cpp
@@ -3,10 +3,10 @@
   *
   **/
 #include "kernel/base/types.h"
-#include "kernel/executors/thread_pool.h"
-#include "kernel/models/reduce.h"
-#include "kernel/models/reduction_operations.h"
-#include "kernel/partitioners/array_partitioner.h"
+#include "kernel/parallel/threading/thread_pool.h"
+#include "kernel/parallel/parallel_algos/parallel_reduce.h"
+#include "kernel/parallel/parallel_algos/reduction_operations.h"
+#include "kernel/parallel/utilities/array_partitioner.h"
 
 #include <thread>
 #include <vector>
