+++ kernel/examples/example_6/exe.cpp
@@ -3,17 +3,18 @@
   *
   **/
 #include "kernel/base/types.h"
-#include "kernel/executors/thread_pool.h"
-#include "kernel/models/reduce.h"
-#include "kernel/models/reduction_operations.h"
-#include "kernel/partitioners/array_partitioner.h"
 #include "kernel/base/algorithm_info.h"
-#include "kernel/data_structs/partitioned_object.h"
-#include "kernel/executors/simple_task.h"
-#include "kernel/models/vector_updater.h"
-#include "kernel/models/scaled_ops.h"
-#include "kernel/models/linear_algebra/dot_product.h"
-#include "kernel/models/linear_algebra/matrix_vector_product.h"
+#include "kernel/parallel/threading/thread_pool.h"
+#include "kernel/parallel/parallel_algos/parallel_reduce.h"
+#include "kernel/parallel/parallel_algos/reduction_operations.h"
+#include "kernel/parallel/utilities/array_partitioner.h"
+#include "kernel/parallel/utilities/partitioned_type.h"
+#include "kernel/parallel/threading/simple_task.h"
+#include "kernel/parallel/parallel_algos/linear_algebra/dot_product.h"
+#include "kernel/parallel/parallel_algos/linear_algebra/matrix_vector_product.h"
+#include "kernel/parallel/parallel_algos/vector_updater.h"
+#include "kernel/utilities/scaled_ops.h"
+
 
 #include <thread>
 #include <vector>
